refactor outcome review handoff routes
